Move Memory GC methods to the new GC file
